[fix]: Fix App Selection

+++ core/network/src/main/java/com/hmh/hamyeonham/core/network/challenge/model/AppsRequest.kt
@@ -13,7 +13,7 @@ data class AppsRequest(
         @SerialName("appCode")
         val appCode: String? = null,
         @SerialName("goalTime")
-        val goalTime: Int? = null
+        val goalTime: Long? = null
     )
 }
